feat: finish DOM structure of Related and child.
