+++ app/Unit/RelatedList.jsx
@@ -7,89 +7,64 @@ import {
 } from 'react-router-dom';
 import {connect} from "react-redux";
 import classnames from 'classnames';
-import styles from "./styles.module.css";
+import SimpleBlock from '../Component/Blocks/SimpleBlock.jsx';
 import {
   handleNounsList,
   handleUsersList
-} from "../../../redux/actions/general.js";
+} from "../redux/actions/general.js";
 import {
   cancelErr,
   uncertainErr
-} from '../../../utils/errHandlers.js';
+} from '../utils/errHandlers.js';
 
 const styleMiddle = {
-  boxFooterInfo: {
-    alignSelf: 'flex-end',
-    margin: '4.2rem 0 1.6rem',
-    padding: '2rem 1.2rem 0',
-    color: '#ababab'
-  },
-  spanFooterInfo: {
-    display: 'inline-block',
-    boxSizing: 'border-block',
-    marginRight: '0.42rem'
-  },
-  textFooterInfo: {
-    fontSize: '1.21rem',
-    letterSpacing: '0.1rem',
-  }
+
 }
 
-class Related extends React.Component {
+class RelatedList extends React.Component {
   constructor(props){
     super(props);
     this.state = {
       axios: false,
-      lastVisit: false,
-      unitsList: [],
+      fetchCount: 0,
+      unitsBlock: [],
+      //unitsBlock is a arr composed of multiple unitsList(also an arr)
       unitsBasic: {},
       marksBasic: {},
     };
     this.axiosSource = axios.CancelToken.source();
-    this._construct_UnitInit = this._construct_UnitInit.bind(this);
-    this._render_IndexNails = this._render_IndexNails.bind(this);
-    this._axios_nouns_singular = this._axios_nouns_singular.bind(this);
+    this._render_Block_relatedList = this._render_Block_relatedList.bind(this);
+    this._axios_GET_relatedFeed = this._axios_GET_relatedFeed.bind(this);
     this.style={
-      withinCom_MainIndex_scroll_: {
-        width: '100%',
-        position: 'absolute',
-        top: '0',
-        left: '0',
-        boxSizing: 'border-box'
-      },
+
     }
   }
 
-  _construct_UnitInit(match, location){
-    let unitInit= {marksify: false, initMark: "all", layer: 0};
-    return unitInit;
-  }
-
-
-  _axios_nouns_singular(){
+  _axios_GET_relatedFeed(){
     const self = this;
     this.setState({axios: true});
 
-    //now get the Units of this noun from the attribution in database
     axios({
       method: 'get',
-      url: '/router/nouns/'+this.nounId,
+      url: '/router/feed/unit/related',
+      params: {unitId: this.props.unitId, fetchCount: this.state.fetchCount},
       headers: {
         'charset': 'utf-8',
         'token': window.localStorage['token']
       },
       cancelToken: self.axiosSource.cancelToken
     }).then(function (res) {
-
       let resObj = JSON.parse(res.data);
+
       self.props._submit_NounsList_new(resObj.main.nounsListMix);
       self.props._submit_UsersList_new(resObj.main.usersList);
       self.setState((prevState, props)=>{
         //we don't push anything and keep it as previous,
-        //bexuase we need to let the render check if there is any id for this noun or not.
+        //because we need to let the render check if there is any id for this noun or not.
         if(resObj.main.unitsList.length>0) prevState.unitsBlock.push(resObj.main.unitsList);
         return({
           axios: false,
+          fetchCount: prevState.fetchCount+1,
           unitsBlock: prevState.unitsBlock, //maybe this is not a good way, modifying the prevState directy
           unitsBasic: Object.assign({}, prevState.unitsBasic, resObj.main.unitsBasic),
           marksBasic: Object.assign({}, prevState.marksBasic, resObj.main.marksBasic)
@@ -106,18 +81,18 @@ class Related extends React.Component {
     });
   }
 
-  _render_nouns_Block(){
+  _render_Block_relatedList(){
+    //before the data fetch or if any empty condition
     if(!this.state.unitsBlock[0]) return(
-      <div
-        style={Object.assign({}, styleMiddle.fontPlaceholder, {boxSizing: 'border-box',margin: '13% 0'})}>
-        {"revealing the unknown to the curious people! "}
+      <div>
+
       </div>
     );
 
     let list = this.state.unitsBlock.map((unitBlock, index)=>{
       return (
         <SimpleBlock
-          key={"key_Cosmicnoun_blocks_"+index}
+          key={"key_block_"+index+"_atRelatedList"}
           unitsList={unitBlock}
           unitsBasic={this.state.unitsBasic}
           marksBasic={this.state.marksBasic}/>
@@ -128,19 +103,15 @@ class Related extends React.Component {
   }
 
   componentDidUpdate(prevProps, prevState, snapshot){
-    //becuase there is chance we jump from noun to noun, using the same component this one
-    //so we check if the nounId has changed
-    this.nounId = this.props.match.params.nounId;
-    if(this.nounId !== prevProps.match.params.nounId){
-      //load Units tagged to this noun
-      this._axios_nouns_singular();
+    //becuase there is chance we jump to another Unit from Related but using the same component
+    //so we check if the unit has changed
+    if(this.props.unitId !== prevProps.unitId){
+      this._axios_GET_relatedFeed();
     };
   }
 
   componentDidMount(){
-    //load Units tagged to this noun
-    this.nounId = this.props.match.params.nounId;
-    this._axios_nouns_singular();
+    this._axios_GET_relatedFeed();
   }
 
   componentWillUnmount(){
@@ -152,27 +123,7 @@ class Related extends React.Component {
   render(){
     return(
       <div>
-        <div
-          style={this.style.withinCom_MainIndex_scroll_}>
-          <div
-            className={classnames(styles.boxTop)}>
-            <div>
-              <RelatedOrigin
-                {...this.props}
-                unitId={this.props.match.params.id}/>
-            </div>
-          </div>
-
-          <div
-            className={styles.boxScroll}>
-            {this._render_IndexNails()}
-          </div>
-
-          <div
-            className={'boxRelativeFull'}
-            style={{height: '156px'}}></div>
-        </div>
-
+        {this._render_Block_relatedList()}
       </div>
     )
   }
@@ -194,4 +145,4 @@ const mapDispatchToProps = (dispatch) => {
 export default withRouter(connect(
   mapStateToProps,
   mapDispatchToProps
-)(Related));
+)(RelatedList));
