+++ app/Unit/RelatedOrigin/styles.module.css
@@ -1,0 +1,33 @@
+.boxNodes {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  width: 17.5%;
+  height: 44%;
+  min-height: 242px;
+  position: absolute;
+  bottom: 4%;
+  left: 0%;
+  box-sizing: border-box;
+  overflow: hidden;
+}
+
+.boxAuthor {
+  color: rgb(250, 250, 250);
+  cursor:pointer;
+}
+
+.boxMark {
+  width: 100%;
+  position: relative;
+  box-sizing: border-box;
+  margin-bottom: 2.8rem;
+}
+
+.boxListRela {
+  width: 100%;
+  box-sizing: border-box;
+  text-align: right;
+  padding-right: 13%;
+  border-right: 0.75px solid #FAFAFA;
+}
