+++ public/css/styleUnitCommon.css
@@ -15,6 +15,15 @@
   color: #FAFAFA;
 }
 
+.nodesListRela{
+  font-size: 2.68rem;
+  letter-spacing: 0.13rem;
+  font-weight: 400;
+  font-family: 'gill-sans-nova', 'Noto Sans TC','Lato', sans-serif;
+  line-height: 2.5rem;
+  color: #FAFAFA;
+}
+
 .fontContentSum {
   font-size: 1.56rem;
   font-weight: 300;
