+++ public/css/stylesWithin.css
@@ -12,6 +12,30 @@
   white-space: nowrap;
   color: #353535;
 }
+._C_C_app_Within_component_LinkExplore_styles_module__comExplore {
+  display: flex;
+  position: relative;
+}
+
+._C_C_app_Within_component_LinkExplore_styles_module__boxExplore {
+  position: relative;
+  box-sizing: border-box;
+}
+
+._C_C_app_Within_component_LinkExplore_styles_module__boxSwitch {
+  display: flex;
+  align-items: flex-end;
+}
+
+._C_C_app_Within_component_LinkExplore_styles_module__boxCategory {
+  display: flex;
+  position: relative;
+  box-sizing: border-box;
+}
+
+._C_C_app_Within_component_LinkExplore_styles_module__fontCategory {
+  color: #a8a8a8;
+}
 ._C_C_app_Within_component_CosmicCorner_stylesCosmicCorner_module__boxOptions {
   box-sizing: border-box;
   text-align: center;
@@ -55,30 +79,6 @@ start the fundemental first, and then the specific condition
   height: 12px;
   right: 186px;
   bottom: 0;
-}
-._C_C_app_Within_component_LinkExplore_styles_module__comExplore {
-  display: flex;
-  position: relative;
-}
-
-._C_C_app_Within_component_LinkExplore_styles_module__boxExplore {
-  position: relative;
-  box-sizing: border-box;
-}
-
-._C_C_app_Within_component_LinkExplore_styles_module__boxSwitch {
-  display: flex;
-  align-items: flex-end;
-}
-
-._C_C_app_Within_component_LinkExplore_styles_module__boxCategory {
-  display: flex;
-  position: relative;
-  box-sizing: border-box;
-}
-
-._C_C_app_Within_component_LinkExplore_styles_module__fontCategory {
-  color: #a8a8a8;
 }
 ._C_C_app_Unit_Unit_styles_module__boxBack {
   display:flex;
@@ -211,6 +211,21 @@ start the fundemental first, and then the specific condition
 ._C_C_app_Unit_Theater_styles_module__spanBackTop {
   box-sizing: border-box;
 }
+._C_C_app_Unit_Related_styles_module__comRelated {
+  width: 100%;
+  position: absolute;
+  top: 0;
+  left: 0;
+  box-sizing: border-box;
+
+}
+
+._C_C_app_Unit_Related_styles_module__boxTop {
+  width: 99%;
+  position: relative;
+  box-sizing: border-box;
+  margin: 0 0 4rem;
+}
 ._C_C_app_Component_Editing_EditingModal_styleEditingModal_module__imgBLockDecoBack {
   width: 18%;
   height: 100%;
@@ -335,6 +350,39 @@ start the fundemental first, and then the specific condition
 
 ._C_C_app_Within_component_MainBanner_styles_module__borderBanner {
   border-bottom: 0.14rem solid #c4c4c4;
+}
+._C_C_app_Unit_RelatedOrigin_styles_module__boxNodes {
+  display: flex;
+  flex-direction: column;
+  justify-content: center;
+  width: 17.5%;
+  height: 44%;
+  min-height: 242px;
+  position: absolute;
+  bottom: 4%;
+  left: 0%;
+  box-sizing: border-box;
+  overflow: hidden;
+}
+
+._C_C_app_Unit_RelatedOrigin_styles_module__boxAuthor {
+  color: rgb(250, 250, 250);
+  cursor:pointer;
+}
+
+._C_C_app_Unit_RelatedOrigin_styles_module__boxMark {
+  width: 100%;
+  position: relative;
+  box-sizing: border-box;
+  margin-bottom: 2.8rem;
+}
+
+._C_C_app_Unit_RelatedOrigin_styles_module__boxListRela {
+  width: 100%;
+  box-sizing: border-box;
+  text-align: right;
+  padding-right: 13%;
+  border-right: 0.75px solid #FAFAFA;
 }
 ._C_C_app_Component_Nails_NailBasic_styleNailBasic_module__Nails_Basic {
   background-color: #FFFFFF;
